% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{fisher_test_with_probability}
\alias{fisher_test_with_probability}
\title{Fisher's Exact Test Analysis}
\usage{
fisher_test_with_probability(observed_table)
}
\arguments{
\item{observed_table}{A contingency table (2x2) of observed counts.}
}
\value{
A list containing:
  \item{FisherTest}{Results of Fisher's exact test.}
  \item{Probability}{The probability of the observed table.}
}
\description{
Fisher's Exact Test Analysis
}
\examples{
observed_table <- matrix(c(2, 8, 0, 0, 0, 1, 5, 3, 1, 0), nrow = 2)
rownames(observed_table) <- c("Smartbond", "Composite")
colnames(observed_table) <- c("1", "2", "3", "4", "5")
fisher_test_with_probability(observed_table)
}
