% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayTable.R
\name{one_way_table_analysis}
\alias{one_way_table_analysis}
\title{One-Way Table Analysis}
\usage{
one_way_table_analysis(data, alpha = 0.05)
}
\arguments{
\item{data}{A matrix representing the one-way table.}

\item{alpha}{The significance level for confidence intervals and chi-squared test.}
}
\value{
A list containing:
  \item{CellProportions}{A matrix of cell proportions.}
  \item{ConfidenceIntervals}{A matrix of confidence intervals for each cell proportion.}
  \item{CellProportionDifferences}{A matrix of differences between cell proportions.}
  \item{ConfidenceIntervalsDifferences}{A matrix of confidence intervals for differences between cell proportions.}
  \item{ChiSquaredTest}{Results of the chi-squared goodness-of-fit test.}
  \item{BarChart}{A bar chart representing the data distribution.}
  \item{bar_chart_chq}{A bar chart representing the chi squared test observe and expected data.}
  \item{pie_chart}{A pie chart representing the data distribution of each category.}
  \item{Summary}{Summary statistics of the one-way table.}
}
\description{
One-Way Table Analysis
}
\examples{
data <- matrix(c(59, 108, 82, 79), nrow = 1)
rownames(data) <- c("Frequency")
colnames(data) <- c("Agree Strongly", "Agree Somewhat", "Disagree Somewhat", "Disagree strongly")
data <- t(data)
rownames(data) <- c("Agree Strongly", "Agree Somewhat", "Disagree Somewhat", "Disagree strongly")
result <- one_way_table_analysis(data, alpha = 0.05)
print(result)
}
