% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoWayTable.R
\name{two_way_table}
\alias{two_way_table}
\title{Two-Way Table Analysis}
\usage{
two_way_table(data, alpha = 0.05)
}
\arguments{
\item{data}{A matrix representing the one-way table.}

\item{alpha}{The significance level for confidence intervals and chi-squared test.}
}
\value{
A list containing:
  \item{ContingencyTable}{The contingency table.}
  \item{ExpectedCounts}{The expected cell counts.}
  \item{ChiSquaredTest}{Results of the chi-squared test.}
  \item{TestStatistic}{The chi-squared test statistic.}
  \item{CriticalValue}{The critical value for chi-squared test.}
  \item{PValue}{The p-value for chi-squared test.}
  \item{RejectRegion}{The rejection region for chi-squared test.}
  \item{Conclusion}{The conclusion based on chi-squared test.}
  \item{Bar chart}{A bar chart representing the data distribution.}
  \item{Mosaicplot}{A Mosaicplot visualize the frequencies in a two way table.}
}
\description{
Two-Way Table Analysis
}
\examples{
conventional_accords <- c(50132 - 5364, 5364)
hybrid_accords <- c(1505 - 137, 137)
data <- matrix(c(conventional_accords, hybrid_accords), nrow = 2)
rownames(data) <- c("Conventional", "Hybrid")
colnames(data) <- c("No Injuries", "Injuries")
two_way_table(data, alpha = 0.05)

}
